chore: move exception package to presentation layer
