+++ src/main/kotlin/smalltalk/backend/infra/repository/room/RoomRepository.kt
@@ -2,6 +2,7 @@ package smalltalk.backend.infra.repository.room
 
 import smalltalk.backend.domain.room.Room
 
+
 interface RoomRepository {
     fun save(roomName: String): Room
     fun getById(roomId: Long): Room
