+++ src/main/kotlin/smalltalk/backend/presentation/exception/room/situation/RoomException.kt
@@ -1,4 +1,4 @@
-package smalltalk.backend.infra.exception.room.situation
+package smalltalk.backend.presentation.exception.room.situation
 
 class RoomIdNotGeneratedException : RuntimeException()
 class RoomNotFoundException : RuntimeException()
